Initial support for --amend, -c and -C commits
As a first step, instead of diffing against the HEAD, we'll diff against
the first parent of whatever SHA1 gets passed in to us. This will work
for the simple case (non-merge commits).

Note that I haven't fleshed out the "strip_submodule_changes" function
as this is starting to get close to the limits of what I feel
comfortable doing in Bash. (I could do some dumb filtering with grep or
similar, but I would prefer to do something a little more rigorous and I
think I need a scripting language to do that without jumping through
hoops.)