Post-merge hook to check submodule changes.
Right now it just informs you that submodules have changed and that it
would be a good idea to run `git submodule update`. This has bitten me
and a few others a few times, and it's not going away.

It would be better to check for updates via grit (ruby git module), but
I'm unconvinced that it would be worth it to install another module,
deal with bundle, etc., to do something that we can do easily anyways.