Try harder to detect end of submodule changes
As we're stripping more aggressively now (since the last commit), try
doubly hard to make sure we know when to stop stripping.

That means stopping not just on truly empty lines, but on any line which
contains only whitespace. I don't expect this to happen much in
practice, but it doesn't hurt to be sure.

Change-Id: I721fb45ab11c4eb1e1a6bc6d3c620f36de290b86
Reviewed-on: https://gerrit.causes.com/571
Tested-by: Greg Hurrell <greg@causes.com>
Reviewed-by: Jimmy Kittiyachavalit <jimmy@causes.com>