Teach `git pair` to echo current Pair info into commit message editor
It's easy to forget to run `git solo` at the end of a pairing session,
and if you then go on to make a commit you might not notice that you and
your pair are still credited as the author/committer.

Normally Git itself will insert an "Author:" annotation in the commit
message editor so that you can realize whenever you're not committing as
yourself, but this won't work for `git pair` because it works by setting
your user.name/user.email directly rather than inserting an `--author`
param to `git commit`.

That was a design decision that we're forced into because we want `git
pair` to work when running multiplexed sessions, and with all manner of
bizarre Git aliases and helper scripts that people use.

So, get git_hooks to insert an "Author:" annotation for us.

This particular hook is probably fairly ripe for refactoring now, as
we're repeating the pattern of running a Ruby snippet over the temporary
file in three different places. If I can think of a nice way to factor
that out I'll follow-up with another commit at some other time.

Related story:
  http://www.pivotaltracker.com/story/show/21808815
  ("Teach `git pair` to echo current Pair info into commit message editor")

Change-Id: Ie0d2b160333696b638eea75ac3ef86b39b1e7414
Reviewed-on: https://gerrit.causes.com/1569
Tested-by: Greg Hurrell <greg@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>