Make hooks more generic
hooks/pre-commit is now a symlink to lib/hook_runner.rb, which
eventually we can make all of the hooks a symlink to. It loads the base
code and then lib/base/#{hook_name}.rb where hook_name is the
underscore-ized version of the hook's name, e.g. 'pre_commit' for the
pre-commit hook.

Next step is to move the individual checks into lib/plugins/

Change-Id: Iead3b001f050eaef0b2a92143afc5cd8bd625acf
Reviewed-on: https://gerrit.causes.com/9798
Tested-by: Aiden Scandella <aiden@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>