Add `user_commit_message` convenience method
To get just the stuff that the user entered, stripping comments and diff
output.

This fixes a bug where `git-commit --verbose` (which `flow` uses) would
potentially have output in the commit message intermediate file which
contained long lines. This should not be a warning for the commit
message, since the user didn't type it.

Change-Id: I3b5223756006fc905538c9c031cacfe47c7c7d3d
Reviewed-on: https://gerrit.causes.com/22239
Tested-by: Aiden Scandella <aiden@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>