First stab at new gem/binary integration
It's not hooked up yet, but we can now build a gem that installs an
`overcommit` binary. That binary uses the newly-created Overcommit::CLI
class, whose implementation is strikingly similar to that of
`scss-lint`. I think Shane did a great job architecting that project, so
I'll be following many of the same patterns.

Change-Id: I9a2d7bc9c042e5c28385bca3647e4c9e4491350f
Reviewed-on: https://gerrit.causes.com/22314
Tested-by: Aiden Scandella <aiden@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>