Add uninstall functionality
Partially because it's the nice thing to do, but also to help clean up
the mess that the current symlink situation has a strangelhold on.

If we just ran the installer in the currently symlinked (old)
installation on developer workstations, we'd end up writing to the
destination of the symlink (~/git_hooks in most cases) rather than the
proper files.

This allows us to first uninstall (via the gem), then install (also via
the gem), then remove git_hooks entirely.

Change-Id: Ibad4ea72e7785294f5163a6fc53c3681636e92ab
Reviewed-on: https://gerrit.causes.com/22843
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Aiden Scandella <aiden@causes.com>