Support native `jshint`
Things have changed since this utility was originally written. The
preferred way of installing/using `jshint` locally is to `npm install -g
jshint`, which puts the binary on your path.

We still support the old rhino/runner way of doing things, but modify
the reporter to match the output format of the `jshint` binary.

Change-Id: I4dfcc5b628df2afd57497304bc59b4e89808a1e8
Reviewed-on: https://gerrit.causes.com/23241
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Aiden Scandella <aiden@causes.com>