Return actual StagedFile objects
This slipped through the cracks when we re-added `StagedFile` in
a7da323. We were forgetting to return an actual `StagedFile` instance,
which lead to any hook that references the `staged` helper of a
`HookSpecificCheck` to fail since it was working with strings instead of
`StagedFile`s.

Also adds a fix that ignores the filetype if no `filetype` is defined
for the check.

Change-Id: Iaf5a8db613299e0d588c8b8d1f4026c06784ee81
Reviewed-on: https://gerrit.causes.com/23949
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>
Tested-by: Greg Hurrell <greg@causes.com>