Always call to_s on output for checks
If a check returned a non-string object for its output, the hook would
fail with an error due to `empty?` not being defined for the object in
question.

To fix this, we explicitly call `to_s` on the output before checking if
it's empty. This also handles the case when there is no output, as
`nil.to_s` is just the empty string.

Change-Id: I1d18e905addc2749f407cf37e1d92dd433d8014b
Reviewed-on: https://gerrit.causes.com/25226
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>