RubyStyle check warns on unexpected output
This check depends on output always matching a regexp that includes a
filename. On other output it would fail attempting to loop up modified
lines for a nil or garbage filename.

Return a warning on unexpected rubocop output instead, printing the
offending output.

This should fix or help debug issue #15.

Change-Id: Ifcc61e5848c9b11271aa6ddc1876f5a4eca761e2
Reviewed-on: https://gerrit.causes.com/27501
Tested-by: jenkins <jenkins@causes.com>
Tested-by: Lann Martin <lann@causes.com>
Reviewed-by: Greg Hurrell <greg@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>