Ignore scss lints on unmodified lines
In order to reduce the noise of `scss-lint` when working with older SCSS
files, we want to change lints on lines you didn't touch into warnings
instead of errors.

This takes the code from 19c009aa almost verbatim and applies it to the
`scss-lint` check. In the interest of time, a more general solution was
not implemented due to a need for this sooner rather than later.

At some point (likely when we add this functionality for the third time)
we'll implement a general solution for this functionality.

Change-Id: I1c9b201ac1186a94e122fbdcf632865debd971d8
Reviewed-on: https://gerrit.causes.com/27569
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Lann Martin <lann@causes.com>