Change Rubocop flags to be 0.12.0-compatible
As of 0.12.0, the `rubocop` CLI no longer requires the `--silent` flag,
and will output a warning if it is used.

Remove the flag from the Ruby style pre-commit check. Note that this
makes this no longer backwards compatible with older versions of
Rubocop, but the 'Unexpected output from rubocop' warning message will
hopefully direct people to upgrade.

Change-Id: Iec239636e367ddfda5fda05920e79981f27ea076
Reviewed-on: https://gerrit.causes.com/28337
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Ryan Fitzgerald <ryan.fitzgerald@causes.com>