Fix rubocop.yml not found in working directory
The switch to using `Pathname#ascend` in e06c0326 broke finding a
`.rubocop.yml` file in the current working directory, since the path
passed to `ascend` would usually be relative (i.e. have no leading `/`),
and so we would not generate a final path of just searching for
`.rubocop.yml`.

Change-Id: I7e66727d077093e3f7c77a03348f488d7e334371
Reviewed-on: https://gerrit.causes.com/30476
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>