Add image_optim as a development dependency
Since we have to `require 'image_optim'` in our tests anyway, add the
`image_optim` gem as a development dependency and remove the class stub.

This still ensures that `overcommit` is decoupled from `image_optim` for
users of Overcommit. For developers, however, it is reasonable to expect
them to have the dependencies of all supported checks so they can run
tests in development.

Also converted some double quotes to single quotes in specs.

Change-Id: I63170434619f5580e3f79ccc7aa6788101e4ca0d
Reviewed-on: https://gerrit.causes.com/30493
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>