Exclude submodules from modified_files
The command to extract modified files currently includes any added
submodule directories. This confuses checks such as the whitespace
checker, which tries to run a `git diff` on them producing messages
such as `bad object :libs/iberty`.

Including submodule directories in the list of modified _files_ doesn't
seem useful, as a submodule isn't a file.

Add the `--ignore-submodules=all` flag to the `git diff` call in
`modified_files` to fix the problem.

Change-Id: Id39414545e57fb0c35e001b7cd90c25c93666f98
Reviewed-on: https://gerrit.causes.com/30773
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>