Add template directory command line switch
Add a switch which results in Overcommit printing the path to a template
directory that can be used to automatically install git hooks any time a
repository is created or cloned.

A simple way to tap into this functionality is to set the
`GIT_TEMPLATE_DIR` environment variable to the result of this call, i.e.

    export GIT_TEMPLATE_DIR=`overcommit --template-dir`

This makes it even easier to use Overcommit for all git projects.

Change-Id: I55aa3f2e33d4f2d476d6e3428c5a8bc57014d301
Reviewed-on: http://gerrit.causes.com/35602
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>