Report error when unsupported Overcommit installed
For users that are already running an older version of Overcommit,
they'll get a weird error if they somehow install the new hooks on their
machine and attempt to run them.

This solves this problem by reporting a helpful error message if the
version of overcommit that was loaded is less than a particular version.

Change-Id: I6fa26c05e0224daffc0d0909bff0d11b91506094
Reviewed-on: http://gerrit.causes.com/35621
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>