Fix plugin loading when custom config defined
There was a problem where if a plugin defined its own custom
configuration Overcommit would attempt to load it from the built-in hook
path instead of the plugin on.

Fix this by differentiating between built-in and plugin hooks in the
`Configuration` class. In the process, refactored the helper methods to
take `HookContext` instead of just a string for the hook type.

Change-Id: I984435676423c2faea2326d226188f0b5c1e86c4
Reviewed-on: http://gerrit.causes.com/35688
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>