Adjust TextWidth subject line default back to 60
Upon further thought, there's no strong reason why a limit of 50
characters is necessary. 72 makes sense for email-based git workflows,
but 50 does not seem to have any strong reason.

Also, the default has been 60 for a long time, so let's keep it.

In the process, renamed some variables and configuration variables for
clarity.

Change-Id: Icd405ecbc3518f00b9741256511b511da6d76f6a
Reviewed-on: http://gerrit.causes.com/35875
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>