Add pre-commit `JsxHint` hook
Folks may want and expect the same amount of linting integration
(lintegration?) when working with JSX files as they have when working
with JavaScript files. This commit adds a pre-commit hook for JsxHint,
which is a wrapper around JSHint to allow linting of files containing
JSX syntax. It claims to be 100% compatible with existing tools using
JSHint.

  https://github.com/STRML/JSXHint

Change-Id: I910ff6110ffa0a640d3fcbf5ed3ba0c999f850dc
Reviewed-on: http://gerrit.causes.com/37618
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Henric Trotzig <henric.trotzig@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>