Allow arguments to passed to ctags in IndexTags
When working in projects with a large number of files, it's nice to be
able to disable generating tags for certain auto-generated code (or
code that is introduced via `npm` or similar mechanism).

Add an option `ctags_arguments` which takes string representing
arguments to pass to the underlying `ctags` executable, allowing the
user to customize how `ctags` is used.

Change-Id: I641c0454497f446f6695d687105927839235b685
Reviewed-on: http://gerrit.causes.com/38070
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>