Skip hard tabs check for golang source files
I've been writing some Go recently, and one of the language's features
is a utility that has an opinion about how your code should be formatted
(like). One of the standards the language has set is that hard tabs
should be used, and it's up to you to configure your editor how to
display.

I think we should skip the hard tabs check by default on Go's source
files for this reason.

Change-Id: I8007004d623fa6bc494ebbbb6bea8747d5511540
Reviewed-on: http://gerrit.causes.com/38560
Reviewed-by: Joe Lencioni <joe.lencioni@causes.com>
Reviewed-by: Shane da Silva <shane@causes.com>
Tested-by: Shane da Silva <shane@causes.com>