Disable IndexTags post-checkout hook by default
This hook doesn't work out of the box on OS X since the version of
`ctags` that ships with XCode is not "exuberant" `ctags`.

Furthermore, since the tags file generated by this hook is typically
only used by developers using the `vim` or `emacs` editor, it isn't
generally applicable enough to be enabled for all developers by default.

Change-Id: I5938a556b2437865bb662d78a53d9487b0aaeb3e
Reviewed-on: http://gerrit.brigade.com/39024
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>