Ignore `json_class` key in JsonSyntax hook
When editing JSON files containing the `json_class` key, the
`JsonSyntax` hook would fail since it used `JSON.load` to parse the
JSON. Since `load` is part of the load/dump interface of `Marshal`, this
would cause problems if the class in the `json_class` key wasn't
available in the runtime environment.

The solution is to use `parse`, which does not give any special
treatment to the `json_class` key.

Change-Id: Ie2a9c3c8b14e87ac494a8ba1e9811740f279213b
Reviewed-on: http://gerrit.causes.com/39263
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>