Fix performance regression in file include/exclude logic
Commit 0e3540d introduced a performance regression where running
Overcommit in repositories with a large number of files that could
potentially apply to a given hook (even if they weren't being staged)
would result in a very slow hook run.

This was due to us switching from `File.fnmatch?` to `Dir.glob` because
at the time the behavior of the latter was what we were trying to
achieve. However, this approach also scans the actual filesystem, making
it slow.

Further investigation revealed that `File.fnmatch?` can take an
additional flag that changes the behavior to act more like `Dir.glob`,
namely resulting in `**/*.rb` matching all Ruby files, both at the root
as well as in subdirectories.

Change-Id: I1e4e66e8aa50d72381301762d5642133515a0b85
Reviewed-on: http://gerrit.causes.com/40419
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>