Add Rubocop to Travis build
Now that the repo has no Rubocop warnings, we can ensure that no more
warnings sneak in by running Rubocop in Travis.

In the process, we pin the version of Rubocop used to 0.24.1, the latest
current version at the time of writing. This will prevent the build from
suddenly failing when a new version of the `rubocop` gem is released.

Change-Id: I831a5f576fa5b89194fcfd054ab8cebebc26fb3f
Reviewed-on: http://gerrit.causes.com/40460
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>