Fix argument passing when overcommit-hook updated
This was a nasty one. In a nutshell, any post-checkout hook would fail
seemingly at random due to command-line arguments not being passed to
it (which should _never_ happen with `post-checkout` hooks).

It turns out that this was caused by the code in `overcommit-hook` which
re-`exec`s the master hook if it was updated before the run. Since we
were passing `$0` (i.e. the master hook script) to `exec` instead of
`$0 $*` (i.e. the master hook script along with all command line
arguments) those arguments were getting lost, causing the post-checkout
hook to fail.

Change-Id: I5a7946c43b772d4a2c0bd68a70e3efc225cd8798
Reviewed-on: http://gerrit.causes.com/40970
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>