Deprecate `:bad` status in favor of `:fail`
This naming convention was a holdover from very early on in Overcommit's
history. However, much of the language used in Overcommit (e.g. the
displayed output from the actual run, or the `fail_hook` matcher in
specs) uses the verb `fail` instead of the adjective `bad`. We would
like to standardize on using verbs going forward.

Since there are currently custom hooks in the wild that are written to
return `:bad`, we'll simply display an deprecation message for now that
warns of the impending removal of this term.

Change-Id: I6cbcce55e04d26bd6215449fb3427b135ab50099
Reviewed-on: http://gerrit.causes.com/40974
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>