Add CLI option to list available hooks
This commit adds a command line option (-l/--list-hooks) to list all the
hooks available for a given repo. All hooks will be printed sorted by
the context in which they run, along with whether the hook is currently
enabled or disabled.

In order to do this I made `Overcommit::Configuration#hook_enabled?`
public. There is logic there to determine whether a certain hook is
enabled or disabled for a given context, and rather than duplicate this
logic elsewhere, it made more sense to make the method available to
`Overcommit::CLI` when printing the hooks.

Change-Id: Ief2bddf04957fb208c50f2faec4b705879825d89
Reviewed-on: http://gerrit.causes.com/42260
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>
Tested-by: Shane da Silva <shane.dasilva@brigade.com>