Don't try to restore modification times for broken symlinks
If a user attempted to commit a broken symlink, the pre-commit setup
step would crash since there would be no file to update after following
the link.

Fix this by testing to see if the file is a broken symlink and ignoring
it. We explicitly don't report an error since it is not the place of
Overcommit to decide if a broken symlink is invalid (that's perhaps
under the purview of a specific hook).

Change-Id: I44e998e50d083be3273fb4b641180f723b583783
Reviewed-on: http://gerrit.causes.com/42395
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>