Restore commit message for cherry-pick conflicts
When cherry-picking a commit onto a branch that resulted in conflicts,
the commit message of the original commit would be lost once the user
resolved those conflicts and attempted to commit.

This was caused by us only saving the `.git/MERGE_MSG` file if
the `MERGE_HEAD` ref was defined. Since `.git/MERGE_MSG` is also used
for cherry-picked commits, we should always attempt to store it.

Change-Id: I0ac4ddc297f4794e50f41256e274cb13e10610e9
Reviewed-on: http://gerrit.causes.com/42836
Tested-by: jenkins <jenkins@causes.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>