Fix installer to work with template dir hooks
When hooks are copied from the template directory specified via
`GIT_TEMPLATE_DIR`, they are copied as files rather than as their
original symlinks.

This resulted in `overcommit --install` not working in repositories
created with these hooks because it didn't handle detecting Overcommit
hooks that weren't symlinks.

Fix this by looking for a special that indicates this is very likely an
overcommit hook (as it's an environment variable that Overcommit
detects).

We don't remove the symlink code since this environment variable was
introduced in 0.17.0, so we want to continue supporting detection via
old methods for a little while longer.

Change-Id: I25fea6a2d20e145ab01963079f676368356a9d81
Reviewed-on: http://gerrit.causes.com/45268
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>