Add command config option to hooks
One feature Overcommit is missing is the ability to execute a command
within a particular environment (e.g. `bundle exec` in Ruby or
`virtualenv` in Python).

In order to support this, add a `command` option which can be configured
in `.overcommit.yml`, and update all existing hooks to respect this
option. If not specified, it defaults to the required executable for the
hook.

Change-Id: Id705739021139babf963ba11a47691acb74e4804
Reviewed-on: http://gerrit.causes.com/46312
Tested-by: jenkins <jenkins@brigade.com>
Reviewed-by: Shane da Silva <shane.dasilva@brigade.com>